Refactor: Now components in problem/ will not refer. that in problem-old/.
