+++ src/app/exam/questions/ExamQuestions.scss
@@ -1,9 +1,5 @@
 
 .exam-questions {
-  .container {
-    padding: 40px 10% 20px 10%;
-  }
-
   .text-center {
     text-align: center;
   }
