+++ src/app/problem-old/ProblemEditorOld.js
@@ -2,16 +2,16 @@ import './ProblemEditorOld.css';
 import {Redirect, useParams} from "react-router-dom";
 import ProblemEditorTitle from "./edit/ProblemEditorTitle";
 import TagList from "./edit/TagList";
-import ProvidedCodeList from './edit/providedCode/ProvidedCodeList';
+import ProvidedCodeList from '../problem/editor/right/ProvidedCodeList';
 import SubmittedCodeList from './edit/SubmittedCodeList';
 import ResourceSpec from "./edit/ResourceSpec";
 import CompilationScript from "./edit/CompilationScript";
 import OutputMatchPolicyList from "./edit/OutputMatchPolicyList";
 import Visible from "./edit/Visible";
-import {EditorButton} from "./edit/EditorButton";
+import {EditorButton} from "../problem/commons/EditorButton";
 import {SubtitleLine} from "../commons/titles/TitleLine";
 import MarkdownEditorOld from "../commons/MarkdownEditorOld";
-import TestCase from "./edit/testCase/TestCasesList";
+import TestCase from "../problem/commons/TestCasesList";
 import React, {useEffect, useState} from "react";
 import {Spinner} from "../commons/Spinner";
 import ProblemNotFound from "./ProblemNotFound";
