+++ src/app/problem-old/edit/TagList.js
@@ -2,7 +2,7 @@ import '../ProblemEditorOld.css';
 import {useTags} from "../../usecases/TagUseCase";
 import {SubtitleLine} from "../../commons/titles/TitleLine";
 import {TextInputField} from "../../commons/TextInputForm/TextInputField";
-import {TextInputItems} from "./TextInputItems";
+import {TextInputItems} from "../../problem/commons/TextInputItems";
 
 function TagList() {
     const {tags, addTag, removeTag} = useTags();
