+++ src/app/problem/commons/EditorButton.js
