+++ src/app/problem/commons/InlineInputBox.js
