+++ src/app/problem/commons/MarkdownEditor.js
@@ -3,16 +3,16 @@ import {useState} from "react";
 
 const MarkdownEditor = ({tabObjects, defaultIndex, isEditing}) => {
     const LEFTMOST_TAB_INDEX = 0;
-    const [currentTabNo, setCurrentTabNo] = useState(LEFTMOST_TAB_INDEX);
+    const [currentTabIndex, setCurrentTabIndex] = useState(LEFTMOST_TAB_INDEX);
 
     return <>
         <div className="markdown">
             {isEditing?
                 <div className="tabs">
                 {tabObjects.map((tabObject, i) =>
-                    <button className={["tab", i === currentTabNo? "current": ""].join(" ")}
+                    <button className={["tab", i === currentTabIndex? "current": ""].join(" ")}
                             style={i !== LEFTMOST_TAB_INDEX? {marginLeft: "-1px"}: {}}
-                            onClick={() => setCurrentTabNo(i)}>
+                            onClick={() => setCurrentTabIndex(i)}>
                         {tabObject.title}
                     </button>
                 )}
@@ -20,7 +20,7 @@ const MarkdownEditor = ({tabObjects, defaultIndex, isEditing}) => {
                 </div>
                 : ""
             }
-            {tabObjects[isEditing? currentTabNo: defaultIndex].component}
+            {tabObjects[isEditing? currentTabIndex: defaultIndex].component}
         </div>
     </>;
 }
