+++ src/app/problem/editor/ProblemEditor.scss
