+++ src/app/problem/editor/left/LeftBar.js
