+++ src/app/problem/editor/left/SectionManager.js
