+++ src/app/problem/editor/left/leftbar.scss
