+++ src/app/problem/editor/right/ResourceSpec.js
@@ -1,6 +1,6 @@
 import Block from "./Block";
 import {useState} from "react";
-import {EditSaveCancelButton} from "../commons/EditSaveCancelButton";
+import {EditSaveCancelButton} from "../../commons/EditSaveCancelButton";
 import {FiCpu} from "react-icons/all";
