+++ src/app/problem/editor/right/RightBar.js
@@ -42,6 +42,7 @@ const RightBar = () => {
             <Divider/>
 
             <TestCases/>
+            <Divider/>
 
             <Actions/>
         </div>
