+++ src/app/problem/editor/right/Tags.js
@@ -1,10 +1,10 @@
 import Block from "./Block";
-import {EditSaveCancelButton} from "../commons/EditSaveCancelButton";
-import {useTags} from "../../usecases/TagUseCase";
-import {TextInputField} from "../../commons/TextInputForm/TextInputField";
-import {TextInputItems} from "../../problem-old/edit/TextInputItems";
+import {EditSaveCancelButton} from "../../commons/EditSaveCancelButton";
+import {useTags} from "../../../usecases/TagUseCase";
+import {TextInputField} from "../../../commons/TextInputForm/TextInputField";
+import {TextInputItems} from "../../commons/TextInputItems";
 import {useState} from "react";
-import TagWithIconList from "../commons/TagWithIconList";
+import TagWithIconList from "../../commons/TagWithIconList";
 import {BsTag} from "react-icons/all";
