+++ src/app/problem/editor/right/Visible.js
@@ -1,4 +1,4 @@
-import {EditorButton} from "../../problem-old/edit/EditorButton";
+import {EditorButton} from "../../commons/EditorButton";
 import Block from "./Block";
 import {useState} from "react";
